+++ ParseSwift.playground/Pages/15 - Custom ObjectId.xcplaygroundpage/Contents.swift
@@ -80,4 +80,36 @@ score.save { result in
     }
 }
 
+//: Fetch object
+score.fetch { result in
+    switch result {
+    case .success(let fetchedScore):
+        print("Successfully fetched: \(fetchedScore)")
+    case .failure(let error):
+        assertionFailure("Error fetching: \(error)")
+    }
+}
+
+//: Query object
+let query = GameScore.query("objectId" == "myObjectId")
+query.first { result in
+    switch result {
+    case .success(let found):
+        print(found)
+    case .failure(let error):
+        print(error)
+    }
+}
+
+//: Query object
+let query = GameScore.query("objectId" == "myObjectId")
+query.find { result in
+    switch result {
+    case .success(let found):
+        print(found)
+    case .failure(let error):
+        print(error)
+    }
+}
+
 //: [Next](@next)