+++ Sources/ParseSwift/Objects/ParseInstallation.swift
@@ -402,7 +402,7 @@ extension ParseInstallation {
     }
 
     func fetchCommand(include: [String]?) throws -> API.Command<Self, Self> {
-        guard isSaved else {
+        guard objectId != nil else {
             throw ParseError(code: .unknownError, message: "Cannot fetch an object without id")
         }